// frontend/src/api/generated/model/maintenanceUrgency.ts
export type MaintenanceUrgency = typeof MaintenanceUrgency[keyof typeof MaintenanceUrgency];


export const MaintenanceUrgency = {
  routine: 'routine',
  urgent: 'urgent',
  emergency: 'emergency',
} as const;